requirements.txt
alembic==1.16.4
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
prometheus_client==0.23.1
prometheus_flask_exporter==0.23.2
pycparser==2.22
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
